refactor: 라우트 가드 로직 통합
