+++ apps/client/src/shared/router/path.ts
@@ -8,4 +8,10 @@ export const PATH = {
   LABEL: '/label/:labelId',
 } as const;
 
+export const GUEST_ONLY_PATHS = [
+  PATH.LANDING,
+  PATH.LOGIN,
+  PATH.LOGIN_CALLBACK,
+] as string[];
+
 export type Routes = (typeof PATH)[keyof typeof PATH];
